+++ baremetal/coremark/core_portme.h
@@ -97,7 +97,7 @@ typedef double ee_f32;
 typedef unsigned char ee_u8;
 typedef unsigned int ee_u32;
 typedef ee_u32 ee_ptr_int;
-typedef size_t ee_size_t;
+typedef uint32_t ee_size_t;
 
 #include "xprintf.h"
 #define ee_printf xprintf
