+++ baremetal/include/trap.h
@@ -2,7 +2,7 @@
 #include "klib-macros.h"
 #include <stdbool.h>
 #include <stdint.h>
-typedef unsigned int size_t;
+typedef unsigned int uint32_t;
 
 inline void check(int cond) {
   volatile int *end_flag = (int *)0x1c;
