+++ baremetal/test/cpu-tests/crc32.c
@@ -6,7 +6,7 @@
 #define STR "The quick brown fox jumps over the lazy dog"
 #define STRLEN (sizeof(STR) - 1)
 
-uint32_t rc_crc32(uint32_t crc, const char *buf, size_t len) {
+uint32_t rc_crc32(uint32_t crc, const char *buf, uint32_t len) {
   static uint32_t table[256];
   static int have_table = 0;
   uint32_t rem;
